Fixed warning in `slave.cpp`.
Removed an implicit cast from `size_t` to `double`.

Review: https://reviews.apache.org/r/52209/

+++ src/slave/slave.hpp
@@ -641,7 +641,7 @@ private:
   // Gauge methods.
   double _frameworks_active()
   {
-    return frameworks.size();
+    return static_cast<double>(frameworks.size());
   }
 
   double _uptime_secs()
